Fix typos&spelling&wording

+++ app/Http/Controllers/OptionalPriorities.php
@@ -22,23 +22,34 @@ class OptionalPriorities extends Controller
 
         $course_id = $request->input('course_id');
         $optionalPLOs = $request->input('optionalItem');
+        
 
-        // Remove option that was not checked
-        Optional_priorities::whereNotIn('custom_PLO',$optionalPLOs)->where('course_id',$course_id)->where('input_status',0)->delete();
+        // Check If Any PLO's have been selected 
+        // Only a Temporary Fix as this doesn't solve the problem when a user wants to unselect PLOS.
+        // when optional_PLOs == null set to empty array
+        if ($optionalPLOs == !NULL) {
 
-        // Loop to insert them to the table
-        foreach($optionalPLOs as $optionalPLO) {
-            if(! (Optional_priorities::where('custom_PLO',$optionalPLO)->where('course_id',$course_id)->first())) {
-                $ops = new Optional_priorities();
-                $ops->course_id = $course_id;
-                $ops->custom_PLO = $optionalPLO;
-                $ops->input_status = 0;
-                if($ops->save()){
-                    $request->session()->flash('success', 'Alignment to UBC/Ministry priorities updated.');
-                }else{
-                    $request->session()->flash('error', 'There was an error updating the alignment to UBC/Ministry priorities.');
+            // ***** Throws error when no PLO's are selected: must be of the type array, null given
+            // Remove option that was not checked
+            Optional_priorities::whereNotIn('custom_PLO',$optionalPLOs)->where('course_id',$course_id)->where('input_status',0)->delete();
+
+            // Loop to insert them to the table
+            foreach($optionalPLOs as $optionalPLO) {
+                if(! (Optional_priorities::where('custom_PLO',$optionalPLO)->where('course_id',$course_id)->first())) {
+                    $ops = new Optional_priorities();
+                    $ops->course_id = $course_id;
+                    $ops->custom_PLO = $optionalPLO;
+                    $ops->input_status = 0;
+                    if($ops->save()){
+                        $request->session()->flash('success', 'Alignment to UBC/Ministry priorities updated.');
+                    }else{
+                        $request->session()->flash('error', 'There was an error updating the alignment to UBC/Ministry priorities.');
+                    }
                 }
             }
+        } else {
+            // Needs to check if PLO's had been previously saved in the db, and if so it need to delete those entries.
+            $request->session()->flash('success', 'Alignment to UBC/Ministry priorities updated.');
         }
 
         /*
